+++ common/src/main/java/me/drex/antixray/common/util/controller/ChunkPacketBlockControllerAntiXray.java
@@ -67,10 +67,7 @@ public abstract class ChunkPacketBlockControllerAntiXray implements ChunkPacketB
         EmptyLevelChunk emptyChunk = new EmptyLevelChunk(level, new ChunkPos(0, 0), level.registryAccess().registryOrThrow(Registries.BIOME).getHolderOrThrow(Biomes.PLAINS));
 
         Block.BLOCK_STATE_REGISTRY.iterator().forEachRemaining((blockState) -> {
-                solidGlobal.put(blockState, blockState.isRedstoneConductor(emptyChunk, BlockPos.ZERO)
-                    && blockState.getBlock() != Blocks.SPAWNER && blockState.getBlock() != Blocks.BARRIER && blockState.getBlock() != Blocks.SHULKER_BOX && blockState.getBlock() != Blocks.SLIME_BLOCK && blockState.getBlock() != Blocks.MANGROVE_ROOTS || lavaObscures && blockState == Blocks.LAVA.defaultBlockState());
-                // Comparing blockState == Blocks.LAVA.defaultBlockState() instead of blockState.getBlock() == Blocks.LAVA ensures that only "stationary lava" is used
-                // shulker box checks TE.
+                solidGlobal.put(blockState, isSolid(emptyChunk, blockState, lavaObscures));
             }
         );
 
@@ -384,6 +381,11 @@ public abstract class ChunkPacketBlockControllerAntiXray implements ChunkPacketB
                 return (int) ((Integer.toUnsignedLong(state) * numberOfBlocks) >>> 32);
             }
         };
+    }
+
+    private static boolean isSolid(EmptyLevelChunk emptyChunk, BlockState blockState, boolean lavaObscures) {
+        return (blockState.isCollisionShapeFullBlock(emptyChunk, BlockPos.ZERO) && blockState.canOcclude()) || (lavaObscures && blockState == Blocks.LAVA.defaultBlockState());
+        // Comparing blockState == Blocks.LAVA.defaultBlockState() instead of blockState.getBlock() == Blocks.LAVA ensures that only "stationary lava" is used
     }
 
 }