+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Enable forge builds
+- Improve solid block check for modded blocks
 
 ## [1.4.2] - 2024-06-17
 ### Fixed